app/requirements.txt
numpy==2.2.1
pandas==2.2.3
streamlit==1.40.2
streamlit_calendar==1.2.1
streamlit-aggrid
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
streamlit-google-auth==1.1.8
